fixed plane generation to envelop wall of tiles of different sizes, cubes are now stretched to cover entire wall, but will be cubic and same sized.
added the saving of uvatlasUVS to a binary file using a rectserializer class.

when models are loaded the tile textures are now checked on the genFlatTexTile component, if it's null then they are reloaded or rebuilt…. to get new textures we need to delete the textures and atlases in the folder, so maybe add a button for this in the iterator…

I would like to move the iterator object to a window, so maybe theres a menu option that creates the generator instance and the iterator object and a window that access of the properties of the iterator object, maybe we can use the same GUI code as on the iterator object…. just mirror it.